git-svn-id: file:///home/danlipsa/dan-svn-repo/src@1455 d596a3c1-de11-4375-84b4-10211c464620

+++ Attribute.cpp
@@ -16,7 +16,7 @@ ostream& IntegerArrayAttribute::Print (ostream& ostr) const
 
 ostream& RealArrayAttribute::Print (ostream& ostr) const
 {
-    ostream_iterator<float> o (ostr, " ");
+    ostream_iterator<double> o (ostr, " ");
     copy (m_values->begin (), m_values->end (), o);
     return ostr;
 }
